Provider: now displays card when not connected
