+++ package.json
@@ -11,6 +11,7 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^13.0.0",
     "@testing-library/user-event": "^13.2.1",
+    "@usedapp/core": "^1.1.5",
     "axios": "^1.1.3",
     "bootstrap": "^5.2.2",
     "classnames": "^2.3.2",
