+++ src/index.js
@@ -8,13 +8,14 @@ import { Provider } from "react-redux"
 
 import store from "./store"
 
-
 const app = (
+
   <Provider store={store}>
     <BrowserRouter>
       <App />
     </BrowserRouter>
   </Provider>
+
 )
 
 ReactDOM.render(app, document.getElementById("root"))
